test: add unit tests for retryWithBackoff
